// lib/aglsl/Destination.d.ts
export declare class Destination {
    mask: number;
    regnum: number;
    regtype: number;
    dim: number;
    indexoffset: number;
    swizzle: number;
    lodbiad: number;
    readmode: number;
    special: number;
    wrap: number;
    filter: number;
    indexregtype: number;
    indexselect: number;
    indirectflag: number;
    constructor();
}